Update V8 to 7b26234c74

+++ deps/v8/src/asmjs/asm-parser.h
@@ -192,7 +192,7 @@ class AsmJsParser {
     }
   }
 
-  inline bool CheckForUnsigned(uint64_t* value) {
+  inline bool CheckForUnsigned(uint32_t* value) {
     if (scanner_.IsUnsigned()) {
       *value = scanner_.AsUnsigned();
       scanner_.Next();
@@ -202,7 +202,7 @@ class AsmJsParser {
     }
   }
 
-  inline bool CheckForUnsignedBelow(uint64_t limit, uint64_t* value) {
+  inline bool CheckForUnsignedBelow(uint32_t limit, uint32_t* value) {
     if (scanner_.IsUnsigned() && scanner_.AsUnsigned() < limit) {
       *value = scanner_.AsUnsigned();
       scanner_.Next();
@@ -294,7 +294,7 @@ class AsmJsParser {
   AsmType* MemberExpression();           // 6.8.5 MemberExpression
   AsmType* AssignmentExpression();       // 6.8.6 AssignmentExpression
   AsmType* UnaryExpression();            // 6.8.7 UnaryExpression
-  AsmType* MultiplicativeExpression();   // 6.8.8 MultaplicativeExpression
+  AsmType* MultiplicativeExpression();   // 6.8.8 MultiplicativeExpression
   AsmType* AdditiveExpression();         // 6.8.9 AdditiveExpression
   AsmType* ShiftExpression();            // 6.8.10 ShiftExpression
   AsmType* RelationalExpression();       // 6.8.11 RelationalExpression
